+++ docs/changelog.rst
@@ -5,4 +5,5 @@ Changelog
 ------------------
 
 - #8 Added patient workflow and managed permissions
+- #7 Added upgrade step handler for senaite.patient
 - Initial Release